assignsubmission_usqfiletypes: implement plugin for type filtering
Adds a submission plugin that, when enabled and configured, limits
the file types the regular 'File submissions' plugin will accept for
upload.

Note that the 'File submissions' plugin must be ordered ahead of this
plugin in the Assignment plugins configuration in order for the file
manager form element to exist before this plugin gets called.