MDL-69362 filelib: fix some quirks of curl::multi()
Fixes reversed logic of the 'CURLOPT_RETURNTRANSFER' option check,
and make request info accessible through get_info() as an array.
Also uses curl_multi_select() within the loop calling multi_exec
to be kind to the CPU if requests are blocking.

I took the liberty of returning the curl_multi error information
in the 'error' and 'errno' properties of the class, and the per-
request error/errno as keys of the 'info' property so as not to
bend the established API too much more.